the most current version
finally was able to parse through the html correctly. just need to set up the server for the email when the price drops to notify the user. Also have to go through and set up the two step authentication so that the server can send an email to the client.